[XLA] Adjust the `CallInliner` pass to allow uniquifying channel ids after inlining.
This is an option that can be enabled specifically on the pass, and is now
enabled in the XLA:GPU compilation pipeline.

This works around an issue whereby the `CallInliner` might inline several
computations involving collectives with the same `channel_id` into the same
computation. This is actually not well-behaved and can result in the creation
of cyclic HLO graphs. This change ensures that inlining does not create cyclic
graphs.

Note that the transformation is actually not semantics-preserving, in that a
computation involving several `call` instructions wrapping collectives that
use the same `channel_id` is already ill-formed.

Nevertheless, `channel_id`s are not used on GPU and are mostly vestigial at
this point. The appropriate solution would be to completely get rid of them,
but they are currently enforced by the HLO verifier and removing them will
take time---hence the workaround.

PiperOrigin-RevId: 737675716